Split computer folder
