+++ computers/__init__.py
@@ -1,5 +1,9 @@
+from .default import *
+from .contrib import *
 from .computer import Computer
-from .browserbase import BrowserbaseBrowser
-from .local_playwright import LocalPlaywrightComputer
-from .docker import DockerComputer
-from .scrapybara import ScrapybaraBrowser, ScrapybaraUbuntu
+
+__all__ = [
+    "default",
+    "contrib",
+    "Computer",
+]
