+++ computers/default/__init__.py
@@ -1,0 +1,12 @@
+from .browserbase import BrowserbaseBrowser
+from .local_playwright import LocalPlaywrightBrowser
+from .docker import DockerComputer
+from .scrapybara import ScrapybaraBrowser, ScrapybaraUbuntu
+
+__all__ = [
+    "BrowserbaseBrowser",
+    "LocalPlaywrightBrowser",
+    "DockerComputer",
+    "ScrapybaraBrowser",
+    "ScrapybaraUbuntu",
+]
