+++ computers/default/browserbase.py
@@ -1,7 +1,7 @@
 import os
 from typing import Tuple, Dict, List, Union, Optional
 from playwright.sync_api import Browser, Page, BrowserContext, Error as PlaywrightError
-from .base_playwright import BasePlaywrightComputer
+from ..shared.base_playwright import BasePlaywrightComputer
 from browserbase import Browserbase
 from dotenv import load_dotenv
 import base64
