+++ computers/default/local_playwright.py
@@ -1,8 +1,8 @@
 from playwright.sync_api import Browser, Page
-from .base_playwright import BasePlaywrightComputer
+from ..shared.base_playwright import BasePlaywrightComputer
 
 
-class LocalPlaywrightComputer(BasePlaywrightComputer):
+class LocalPlaywrightBrowser(BasePlaywrightComputer):
     """Launches a local Chromium instance using Playwright."""
 
     def __init__(self, headless: bool = False):
@@ -11,12 +11,13 @@ class LocalPlaywrightComputer(BasePlaywrightComputer):
 
     def _get_browser_and_page(self) -> tuple[Browser, Page]:
         width, height = self.dimensions
-        launch_args = [f"--window-size={width},{height}", "--disable-extensions", "--disable-file-system"]
+        launch_args = [
+            f"--window-size={width},{height}",
+            "--disable-extensions",
+            "--disable-file-system",
+        ]
         browser = self._playwright.chromium.launch(
-            chromium_sandbox=True,
-            headless=self.headless,
-            args=launch_args,
-            env={}
+            chromium_sandbox=True, headless=self.headless, args=launch_args, env={}
         )
         page = browser.new_page()
         page.set_viewport_size({"width": width, "height": height})
